knowledge_base/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
langchain>=0.0.300
langchain-community>=0.0.20
langchain-openai>=0.0.5
chromadb>=0.4.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0,<0.10.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
tqdm>=4.60.0
pandas>=2.0.0
numpy>=1.24.0
